// apps/frontend/web/src/types/points.ts
export type PointsTab = 'history' | 'referrals';

export type BadgeType = 'weekly_3_orders' | 'weekly_5_orders';

export interface ReferralItem {
  id: number;
  referred_user: {
    id: number;
    name: string;
    username: string;
    profile_image_url: null | string;
  };
  status: 'pending' | 'completed';
  completed_at: null | string;
  created_at: string;
}

export interface WalletDetails {
  id: number;
  user_id: number;
  balance: number;
  balance_in_egp: number;
  created_at: string;
  updated_at: string;
}

export interface StreakDetails {
  week_start_date: string;
  completed_orders_count: number;
  next_tier: null | {
    target_orders: number;
    orders_needed: number;
    reward_points: number;
    badge_type: BadgeType | null;
  };
  badges:
    | {
        id: number;
        badge_type: BadgeType;
        week_start_date: string;
        created_at: string;
      }[]
    | [];
}

export interface WalletHistory {
  id: number;
  points: number;
  type: 'earn' | 'redeem' | 'gift_sent' | 'gift_received';
  source:
    | 'referral'
    | 'redemption'
    | 'leaderboard'
    | 'weekly_streak'
    | 'point_gift';
  description: string | null;
  reference_id: number;
  created_at: string;
}